practice_mongodb/pyproject.toml
[project]
name = "practice-mongodb"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pymongo[srv]>=4.12.0",
    "python-dotenv>=1.1.0",
    "rich>=14.0.0",
]
